--- a/.gitignore
+++ b/.gitignore
@@ -196,7 +196,7 @@ Xcode/
 /build-dir/
 
 # precompiled headers
-/pch/pch.hpp.gch
-/pch/pch.hpp.pch
-/tests/pch/pch.hpp.gch
-/tests/pch/pch.hpp.pch
+/pch/main-pch.hpp.gch
+/pch/main-pch.hpp.pch
+/tests/pch/tests-pch.hpp.gch
+/tests/pch/tests-pch.hpp.pch