--- a/Makefile
+++ b/Makefile
@@ -163,6 +163,10 @@ ifndef RUNTESTS
   RUNTESTS = 1
 endif
 
+ifndef PCH
+  PCH = 1
+endif
+
 # Auto-detect MSYS2
 ifdef MSYSTEM
   MSYS2 = 1
@@ -253,6 +257,12 @@ ifneq ($(findstring BSD,$(OS)),)
   BSD = 1
 endif
 
+ifeq ($(PCH), 1)
+	CCACHEBIN = CCACHE_SLOPPINESS=pch_defines,time_macros ccache
+else
+	CCACHEBIN = ccache
+endif
+
 # This sets CXX and so must be up here
 ifneq ($(CLANG), 0)
   ifeq ($(NATIVE), osx)
@@ -267,8 +277,8 @@ ifneq ($(CLANG), 0)
     CXXFLAGS += -fno-builtin
   endif
   ifeq ($(CCACHE), 1)
-    CXX = CCACHE_CPP2=1 ccache $(CROSS)$(COMPILER)
-    LD  = CCACHE_CPP2=1 ccache $(CROSS)$(COMPILER)
+    CXX = CCACHE_CPP2=1 $(CCACHEBIN) $(CROSS)$(COMPILER)
+    LD  = CCACHE_CPP2=1 $(CCACHEBIN) $(CROSS)$(COMPILER)
   else
     CXX = $(CROSS)$(COMPILER)
     LD  = $(CROSS)$(COMPILER)
@@ -281,8 +291,8 @@ else
   endif
 
   ifeq ($(CCACHE), 1)
-    CXX = ccache $(CROSS)$(COMPILER)
-    LD  = ccache $(CROSS)$(COMPILER)
+    CXX = $(CCACHEBIN) $(CROSS)$(COMPILER)
+    LD  = $(CCACHEBIN) $(CROSS)$(COMPILER)
   else
     CXX = $(CROSS)$(COMPILER)
     LD  = $(CROSS)$(COMPILER)
@@ -400,6 +410,35 @@ endif
 
 ifeq ($(CYGWIN),1)
 WARNINGS += -Wimplicit-fallthrough=0
+endif
+
+ifeq ($(PCH), 1)
+  PCHFLAGS = -Ipch -Winvalid-pch
+  PCH_H = pch/pch.hpp
+
+  ifeq ($(CLANG), 0)
+    PCHFLAGS += -fpch-preprocess -include pch.hpp
+    PCH_P = pch/pch.hpp.gch
+  else
+    PCH_P = pch/pch.hpp.pch
+    PCHFLAGS += -include-pch $(PCH_P)
+
+    # FIXME: dirty hack ahead
+    # ccache won't wort with clang unless it supports -fno-pch-timestamp
+    ifeq ($(CCACHE), 1)
+      CLANGVER := $(shell echo 'int main(void){return 0;}'|$(CXX) -Xclang -fno-pch-timestamp -x c++ -o $(ODIR)/__bla__.o - 2>&1)
+      ifneq ($(.SHELLSTATUS), 0)
+        PCHFLAGS = ""
+        PCH_H = ""
+        PCH_P = ""
+        PCH = 0
+        $(warning your clang version does not support -fno-pch-timestamp: $(CLANGVER) ($(.SHELLSTATUS)))
+      else
+        CXXFLAGS += -Xclang -fno-pch-timestamp
+      endif
+    endif
+    
+  endif
 endif
 
 CXXFLAGS += $(WARNINGS) $(DEBUG) $(DEBUGSYMS) $(PROFILE) $(OTHERS) -MMD -MP
@@ -827,6 +866,9 @@ ifeq ($(RELEASE), 1)
   endif
 endif
 
+$(PCH_P): $(PCH_H)
+	-$(CXX) $(CPPFLAGS) $(DEFINES) $(subst -Werror,,$(CXXFLAGS)) -c $(PCH_H) -o $(PCH_P)
+
 $(BUILD_PREFIX)$(TARGET_NAME).a: $(OBJS)
 	$(AR) rcs $(BUILD_PREFIX)$(TARGET_NAME).a $(filter-out $(ODIR)/main.o $(ODIR)/messages.o,$(OBJS))
 
@@ -841,8 +883,8 @@ version:
 # Unconditionally create the object dir on every invocation.
 $(shell mkdir -p $(ODIR))
 
-$(ODIR)/%.o: $(SRC_DIR)/%.cpp
-	$(CXX) $(CPPFLAGS) $(DEFINES) $(CXXFLAGS) -c $< -o $@
+$(ODIR)/%.o: $(SRC_DIR)/%.cpp $(PCH_P)
+	$(CXX) $(CPPFLAGS) $(DEFINES) $(CXXFLAGS) $(PCHFLAGS) -c $< -o $@
 
 $(ODIR)/%.o: $(SRC_DIR)/%.rc
 	$(RC) $(RFLAGS) $< -o $@
@@ -867,6 +909,7 @@ clean: clean-tests
 	rm -rf *$(BINDIST_DIR) *cataclysmbn-*.tar.gz *cataclysmbn-*.zip
 	rm -f $(SRC_DIR)/version.h
 	rm -f $(CHKJSON_BIN)
+	rm -f pch/pch.hpp.{gch,pch}
 
 distclean:
 	rm -rf *$(BINDIST_DIR)
@@ -1043,7 +1086,7 @@ ifdef LANGUAGES
 endif
 	$(BINDIST_CMD)
 
-export ODIR _OBJS LDFLAGS CXX W32FLAGS DEFINES CXXFLAGS TARGETSYSTEM
+export ODIR _OBJS LDFLAGS CXX W32FLAGS DEFINES CXXFLAGS TARGETSYSTEM CLANG PCH
 
 ctags: $(ASTYLE_SOURCES)
 	ctags $^