--- a/Makefile
+++ b/Makefile
@@ -909,7 +909,7 @@ clean: clean-tests
 	rm -rf *$(BINDIST_DIR) *cataclysmbn-*.tar.gz *cataclysmbn-*.zip
 	rm -f $(SRC_DIR)/version.h
 	rm -f $(CHKJSON_BIN)
-	rm -f pch/pch.hpp.{gch,pch}
+	rm -f pch/*pch.hpp.{gch,pch,d}
 
 distclean:
 	rm -rf *$(BINDIST_DIR)