--- a/tests/Makefile
+++ b/tests/Makefile
@@ -27,6 +27,43 @@ CXXFLAGS += -I../src -Wno-unused-variable -Wno-sign-compare -Wno-unknown-pragmas
 CXXFLAGS += -Wall -Wextra \
   -Wno-range-loop-analysis # TODO: Fix warnings instead of disabling
 
+ifndef PCH
+  PCH = 1
+endif
+
+ifndef CLANG
+  CLANG = 0
+endif
+
+ifeq ($(PCH), 1)
+  PCHFLAGS = -Ipch -Winvalid-pch -DCATA_CATCH_PCH
+  PCH_H = pch/pch.hpp
+  ifeq ($(CLANG), 0)
+    PCH_P = pch/pch.hpp.gch
+    PCHFLAGS += -fpch-preprocess -include pch.hpp
+  else
+    PCH_P = pch/pch.hpp.pch
+    PCHFLAGS += -include-pch $(PCH_P)
+    CXXFLAGS += -Wno-unused-macros
+    
+    # FIXME: dirty hack ahead
+    # ccache won't wort with clang unless it supports -fno-pch-timestamp
+    ifeq ($(CCACHE), 1)
+      CLANGVER := $(shell echo 'int main(void){return 0;}'|$(CXX) -Xclang -fno-pch-timestamp -x c++ -o $(ODIR)/__bla__.o - 2>&1)
+      ifneq ($(.SHELLSTATUS), 0)
+        $(warning your clang version does not support -fno-pch-timestamp: $(CLANGVER) ($(.SHELLSTATUS)))
+        undefine PCHFLAGS
+        undefine PCH_H
+        undefine PCH_P
+        PCH = 0
+      else
+        CXXFLAGS += -Xclang -fno-pch-timestamp
+      endif
+    endif
+        
+  endif
+endif
+
 ifeq ($(TARGETSYSTEM), WINDOWS)
   TEST_TARGET = $(BUILD_PREFIX)cata_test.exe
 else
@@ -35,11 +72,11 @@ endif
 
 tests: $(TEST_TARGET)
 
-$(TEST_TARGET): $(OBJS) $(CATA_LIB) precompile_header
-	+$(CXX) $(W32FLAGS) -o $@ $(DEFINES) $(OBJS) $(CATA_LIB) $(CXXFLAGS) $(LDFLAGS)
+$(TEST_TARGET): $(OBJS) $(CATA_LIB)
+	+$(CXX) $(W32FLAGS) -o $@ $(DEFINES) $(OBJS) $(CATA_LIB) $(CXXFLAGS) $(PCHFLAGS) $(LDFLAGS)
 
-precompile_header:
-	-$(CXX) $(CPPFLAGS) $(DEFINES) $(subst -Werror,,$(CXXFLAGS)) -Wno-non-virtual-dtor -Wno-unused-macros -DCATCH_CONFIG_ENABLE_PAIR_STRINGMAKER -c catch/catch.hpp -o catch/catch.hpp.gch
+$(PCH_P): $(PCH_H)
+	-$(CXX) $(CPPFLAGS) $(DEFINES) $(subst -Werror,,$(CXXFLAGS)) -Wno-non-virtual-dtor -Wno-unused-macros -I. -c $(PCH_H) -o $(PCH_P)
 
 # Iterate over all the individual tests.
 check: $(TEST_TARGET)
@@ -48,12 +85,13 @@ check: $(TEST_TARGET)
 clean:
 	rm -rf *obj *objwin
 	rm -f *cata_test
+	rm -f pch/pch.hpp.{gch,pch}
 
 #Unconditionally create object directory on invocation.
 $(shell mkdir -p $(ODIR))
 
-$(ODIR)/%.o: %.cpp
-	$(CXX) $(CPPFLAGS) $(DEFINES) -DCATCH_CONFIG_ENABLE_PAIR_STRINGMAKER $(CXXFLAGS) -c $< -o $@
+$(ODIR)/%.o: %.cpp $(PCH_P)
+	$(CXX) $(CPPFLAGS) $(DEFINES) $(CXXFLAGS) $(PCHFLAGS) -c $< -o $@
 
 .PHONY: clean check tests precompile_header