--- a/tests/Makefile
+++ b/tests/Makefile
@@ -36,32 +36,14 @@ ifndef CLANG
 endif
 
 ifeq ($(PCH), 1)
-  PCHFLAGS = -Ipch -Winvalid-pch -DCATA_CATCH_PCH
-  PCH_H = pch/pch.hpp
-  ifeq ($(CLANG), 0)
-    PCH_P = pch/pch.hpp.gch
-    PCHFLAGS += -fpch-preprocess -include pch.hpp
-  else
-    PCH_P = pch/pch.hpp.pch
-    PCHFLAGS += -include-pch $(PCH_P)
-    CXXFLAGS += -Wno-unused-macros
-    
-    # FIXME: dirty hack ahead
-    # ccache won't wort with clang unless it supports -fno-pch-timestamp
-    ifeq ($(CCACHE), 1)
-      CLANGVER := $(shell echo 'int main(void){return 0;}'|$(CXX) -Xclang -fno-pch-timestamp -x c++ -o $(ODIR)/__bla__.o - 2>&1)
-      ifneq ($(.SHELLSTATUS), 0)
-        $(warning your clang version does not support -fno-pch-timestamp: $(CLANGVER) ($(.SHELLSTATUS)))
-        undefine PCHFLAGS
-        undefine PCH_H
-        undefine PCH_P
-        PCH = 0
-      else
-        CXXFLAGS += -Xclang -fno-pch-timestamp
-      endif
-    endif
-        
-  endif
+PCHFLAGS += -DCATA_CATCH_PCH
+PCH_H = pch/tests-pch.hpp
+ifeq ($(CLANG), 0)
+PCH_P = $(PCH_H).gch
+else
+PCH_P = $(PCH_H).pch
+CXXFLAGS += -Wno-unused-macros
+endif
 endif
 
 ifeq ($(TARGETSYSTEM), WINDOWS)
@@ -73,7 +55,7 @@ endif
 tests: $(TEST_TARGET)
 
 $(TEST_TARGET): $(OBJS) $(CATA_LIB)
-	+$(CXX) $(W32FLAGS) -o $@ $(DEFINES) $(OBJS) $(CATA_LIB) $(CXXFLAGS) $(PCHFLAGS) $(LDFLAGS)
+	+$(CXX) $(W32FLAGS) -o $@ $(DEFINES) $(OBJS) $(CATA_LIB) $(CXXFLAGS) $(LDFLAGS)
 
 $(PCH_P): $(PCH_H)
 	-$(CXX) $(CPPFLAGS) $(DEFINES) $(subst -Werror,,$(CXXFLAGS)) -Wno-non-virtual-dtor -Wno-unused-macros -I. -c $(PCH_H) -o $(PCH_P)
@@ -91,7 +73,7 @@ clean:
 $(shell mkdir -p $(ODIR))
 
 $(ODIR)/%.o: %.cpp $(PCH_P)
-	$(CXX) $(CPPFLAGS) $(DEFINES) $(CXXFLAGS) $(PCHFLAGS) -c $< -o $@
+	$(CXX) $(CPPFLAGS) $(DEFINES) $(CXXFLAGS) $(subst main-pch,tests-pch,$(PCHFLAGS)) -c $< -o $@
 
 .PHONY: clean check tests precompile_header