--- a/tests/Makefile
+++ b/tests/Makefile
@@ -67,7 +67,7 @@ check: $(TEST_TARGET)
 clean:
 	rm -rf *obj *objwin
 	rm -f *cata_test
-	rm -f pch/pch.hpp.{gch,pch}
+	rm -f pch/*pch.hpp.{gch,pch,d}
 
 #Unconditionally create object directory on invocation.
 $(shell mkdir -p $(ODIR))