--- a/tests/Makefile
+++ b/tests/Makefile
@@ -67,7 +67,9 @@ check: $(TEST_TARGET)
 clean:
 	rm -rf *obj *objwin
 	rm -f *cata_test
-	rm -f pch/*pch.hpp.{gch,pch,d}
+	rm -f pch/*pch.hpp.gch
+	rm -f pch/*pch.hpp.pch
+	rm -f pch/*pch.hpp.d
 
 #Unconditionally create object directory on invocation.
 $(shell mkdir -p $(ODIR))