--- a/tests/active_item_cache_test.cpp
+++ b/tests/active_item_cache_test.cpp
@@ -1,8 +1,9 @@
+#include "catch/catch.hpp"
+
 #include <memory>
 #include <set>
 
 #include "calendar.h"
-#include "catch/catch.hpp"
 #include "game.h"
 #include "game_constants.h"
 #include "item.h"