--- a/tests/behavior_test.cpp
+++ b/tests/behavior_test.cpp
@@ -1,10 +1,11 @@
+#include "catch/catch.hpp"
+
 #include <memory>
 #include <string>
 #include <vector>
 
 #include "behavior.h"
 #include "behavior_strategy.h"
-#include "catch/catch.hpp"
 #include "character_oracle.h"
 #include "game.h"
 #include "item.h"