--- a/tests/bionics_test.cpp
+++ b/tests/bionics_test.cpp
@@ -1,3 +1,5 @@
+#include "catch/catch.hpp"
+
 #include <climits>
 #include <list>
 #include <memory>
@@ -6,7 +8,6 @@
 #include "avatar.h"
 #include "bionics.h"
 #include "calendar.h"
-#include "catch/catch.hpp"
 #include "item.h"
 #include "pimpl.h"
 #include "player.h"