--- a/tests/cata_utility_test.cpp
+++ b/tests/cata_utility_test.cpp
@@ -1,4 +1,5 @@
 #include "catch/catch.hpp"
+
 #include "assertion_helpers.h"
 #include "cata_utility.h"
 #include "string_utils.h"