--- a/tests/char_edible_rating_test.cpp
+++ b/tests/char_edible_rating_test.cpp
@@ -1,9 +1,10 @@
+#include "catch/catch.hpp"
+
 #include <memory>
 #include <string>
 
 #include "avatar.h"
 #include "calendar.h"
-#include "catch/catch.hpp"
 #include "character.h"
 #include "item.h"
 #include "itype.h"