--- a/tests/char_healing_test.cpp
+++ b/tests/char_healing_test.cpp
@@ -1,3 +1,5 @@
+#include "catch/catch.hpp"
+
 #include <string>
 
 #include "avatar.h"