--- a/tests/char_validity_check_test.cpp
+++ b/tests/char_validity_check_test.cpp
@@ -1,4 +1,5 @@
 #include "catch/catch.hpp"
+
 #include "char_validity_check.h"
 
 TEST_CASE( "char_validity_check" )