--- a/tests/colony_test.cpp
+++ b/tests/colony_test.cpp
@@ -1,10 +1,11 @@
+#include "catch/catch.hpp"
+
 #include <cstddef>
 #include <algorithm> // std::find
 #include <functional> // std::greater
 #include <utility> // std::move
 #include <vector> // range-insert testing
 
-#include "catch/catch.hpp"
 #include "colony.h"
 #include "colony_list_test_helpers.h"