--- a/tests/comestible_test.cpp
+++ b/tests/comestible_test.cpp
@@ -1,3 +1,5 @@
+#include "catch/catch.hpp"
+
 #include <algorithm>
 #include <cstdio>
 #include <memory>
@@ -5,7 +7,6 @@
 #include <vector>
 
 #include "character.h"
-#include "catch/catch.hpp"
 #include "item.h"
 #include "item_contents.h"
 #include "itype.h"