--- a/tests/creature_in_field_test.cpp
+++ b/tests/creature_in_field_test.cpp
@@ -1,6 +1,7 @@
+#include "catch/catch.hpp"
+
 #include <memory>
 
-#include "catch/catch.hpp"
 #include "game.h"
 #include "map.h"
 #include "map_helpers.h"