--- a/tests/creature_test.cpp
+++ b/tests/creature_test.cpp
@@ -1,8 +1,9 @@
+#include "catch/catch.hpp"
+
 #include <cstdlib>
 #include <map>
 #include <utility>
 
-#include "catch/catch.hpp"
 #include "creature.h"
 #include "monster.h"
 #include "mtype.h"