--- a/tests/dependency_tree_test.cpp
+++ b/tests/dependency_tree_test.cpp
@@ -1,6 +1,6 @@
 #include "catch/catch.hpp"
-#include "stringmaker.h"
 
+#include "stringmaker.h"
 #include "mod_manager.h"
 #include "dependency_tree.h"