--- a/tests/event_test.cpp
+++ b/tests/event_test.cpp
@@ -1,10 +1,11 @@
+#include "catch/catch.hpp"
+
 #include <algorithm>
 #include <string>
 #include <vector>
 
 #include "calendar.h"
 #include "cata_variant.h"
-#include "catch/catch.hpp"
 #include "character_id.h"
 #include "event.h"
 #include "event_bus.h"