--- a/tests/fold_string_test.cpp
+++ b/tests/fold_string_test.cpp
@@ -1,9 +1,10 @@
+#include "catch/catch.hpp"
+
 #include <initializer_list>
 #include <iterator>
 #include <vector>
 #include <string>
 
-#include "catch/catch.hpp"
 #include "output.h"
 
 template<class IterResult, class IterExpect>