--- a/tests/health_test.cpp
+++ b/tests/health_test.cpp
@@ -1,3 +1,5 @@
+#include "catch/catch.hpp"
+
 #include <algorithm>
 #include <array>
 #include <cstddef>
@@ -5,7 +7,6 @@
 #include <vector>
 
 #include "calendar.h"
-#include "catch/catch.hpp"
 #include "enums.h"
 #include "npc.h"