--- a/tests/item_group_test.cpp
+++ b/tests/item_group_test.cpp
@@ -4,6 +4,7 @@
 #include <vector>
 
 #include "catch/catch.hpp"
+
 #include "item.h"
 #include "item_group.h"
 #include "stringmaker.h"