--- a/tests/item_test.cpp
+++ b/tests/item_test.cpp
@@ -1,9 +1,10 @@
+#include "catch/catch.hpp"
+
 #include <initializer_list>
 #include <limits>
 #include <memory>
 
 #include "calendar.h"
-#include "catch/catch.hpp"
 #include "enums.h"
 #include "item.h"
 #include "itype.h"