--- a/tests/line_test.cpp
+++ b/tests/line_test.cpp
@@ -1,3 +1,5 @@
+#include "catch/catch.hpp"
+
 #include <algorithm>
 #include <chrono>
 #include <cstdio>
@@ -8,7 +10,6 @@
 #include <vector>
 
 #include "cata_generators.h"
-#include "catch/catch.hpp"
 #include "coordinates.h"
 #include "line.h"
 #include "point.h"