--- a/tests/magic_spell_test.cpp
+++ b/tests/magic_spell_test.cpp
@@ -1,9 +1,10 @@
+#include "catch/catch.hpp"
+
 #include "avatar.h"
 #include "game.h"
 #include "magic.h"
 #include "monster.h"
 
-#include "catch/catch.hpp"
 #include "player_helpers.h"
 #include "map_helpers.h"