--- a/tests/map_iterator_test.cpp
+++ b/tests/map_iterator_test.cpp
@@ -1,7 +1,8 @@
+#include "catch/catch.hpp"
+
 #include <algorithm>
 #include <array>
 
-#include "catch/catch.hpp"
 #include "coordinates.h"
 #include "map_iterator.h"
 #include "point.h"