--- a/tests/map_test.cpp
+++ b/tests/map_test.cpp
@@ -1,8 +1,9 @@
+#include "catch/catch.hpp"
+
 #include <memory>
 #include <vector>
 
 #include "avatar.h"
-#include "catch/catch.hpp"
 #include "enums.h"
 #include "game.h"
 #include "game_constants.h"