--- a/tests/monster_vision_test.cpp
+++ b/tests/monster_vision_test.cpp
@@ -1,7 +1,8 @@
+#include "catch/catch.hpp"
+
 #include <memory>
 
 #include "calendar.h"
-#include "catch/catch.hpp"
 #include "game.h"
 #include "map.h"
 #include "map_helpers.h"