--- a/tests/pch/pch.hpp
+++ b/tests/pch/pch.hpp
@@ -1,0 +1,2 @@
+#define CATCH_CONFIG_ALL_PARTS
+#include "catch/catch.hpp"