--- a/update-pch.sh
+++ b/update-pch.sh
@@ -3,7 +3,7 @@
 suffix=`date|md5sum|awk '{print $1}'`
 tpchr=/tmp/_pchr_$suffix.hpp
 tpch=/tmp/_pch_$suffix.hpp
-pch=pch/pch.hpp
+pch=pch/main-pch.hpp
 
 grep '#include <' -R src|grep -v 'NOPCH'|awk -F '[: ]' '{print $2 " " $3}' > $tpchr
 grep -v '\.h' $tpchr|sort|uniq >> $tpch