#### Summary
SUMMARY: Infrastructure "[DDA Port] Use precompiled headers to speed up Make build"

#### Purpose of change
Build speedup

#### Describe the solution
Cherry-picked the following PRs:
https://github.com/CleverRaven/Cataclysm-DDA/pull/42453
https://github.com/CleverRaven/Cataclysm-DDA/pull/42535
https://github.com/CleverRaven/Cataclysm-DDA/pull/42600
https://github.com/CleverRaven/Cataclysm-DDA/pull/42767
https://github.com/CleverRaven/Cataclysm-DDA/pull/43080
https://github.com/CleverRaven/Cataclysm-DDA/pull/43194
https://github.com/CleverRaven/Cataclysm-DDA/pull/43291

#### Describe alternatives you've considered
Painting my laptop red

#### Testing
Compiled with `time make -j4 -k COMPILER=clang++-12 TILES=1`

Before:
```
real    22m7,844s
user    83m55,413s
sys     1m56,165s
```

After:
```
real    20m22,521s
user    75m19,058s
sys     1m47,387s
```